allow existing macro type params to be valid
